requirements.txt
# MedSafe AI - Python Dependencies

# Frontend
streamlit>=1.44.0
requests>=2.32.3

# Backend API
fastapi>=0.115.0
uvicorn>=0.30.6
python-multipart>=0.0.9

# OCR and Image Processing
pytesseract>=0.3.10
Pillow>=11.0.0
numpy>=1.26.0
scipy>=1.13.0
scikit-image>=0.24.0

# Fuzzy Matching
rapidfuzz>=3.6.1

# LLM Integration
ollama>=0.4.0

# Data Handling
pandas>=2.2.0

# Utilities
python-dateutil>=2.8.2
psutil>=6.0.0

# Testing (optional)
pytest>=8.0.0
pytest-cov>=4.1.0
